Change name file to button-heart-icon-words
