+++ app/views/kitten/components/buttons/button-heart-icon-words.html.erb
@@ -1,39 +1,39 @@
 <div class="karl-Example__column">
   <%= example "Hydrogen button with icon and words" do %>
-    <%= react_component('ButtonIconWords') %>
+    <%= react_component('ButtonHeartIconWords') %>
   <% end %>
 
 
   <%= example "Helium button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         name: 'helium',
       }
     ) %>
   <% end %>
 
   <%= example "Lithium button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         name: 'lithium',
       }
     ) %>
   <% end %>
 
   <%= example "Beryllium button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         name: 'beryllium',
       }
     ) %>
   <% end %>
 
   <%= example "Boron button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         name: 'boron',
       }
     ) %>
   <% end %>
 
   <%= example "Carbon button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         name: 'carbon',
       }
     ) %>
@@ -42,7 +42,7 @@
 
 <div class="karl-Example__column">
   <%= example "Disabled button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         disabled: true,
       }
     ) %>
@@ -51,7 +51,7 @@
 
 <div class="karl-Example__column">
   <%= example "Tiny button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         size: 'tiny',
         text: 'btn',
       }
@@ -59,7 +59,7 @@
   <% end %>
 
   <%= example "Tiny button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         size: 'tiny',
         text: 'btn',
         iconOnRight: true,
@@ -68,7 +68,7 @@
   <% end %>
 
   <%= example "Big button with icon and words" do %>
-    <%= react_component('ButtonIconWords', props: {
+    <%= react_component('ButtonHeartIconWords', props: {
         size: 'big',
       }
     ) %>
