+++ assets/javascripts/kitten/components/buttons/button-heart-icon-words.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import Button from './button'
+import { Button } from 'kitten/components/buttons/button'
 import { HeartIcon } from 'kitten/components/icons/heart-icon'
 
 export class ButtonHeartIconWords extends React.Component {
   render() {
     return (
-      <Button { ...this.props } className='k-Button--icon' >
+      <Button { ...this.props } className='k-Button--icon'>
         <HeartIcon className="k-Button__icon" />
       </Button>
     )
